requirements.txt
streamlit>=1.36
pandas
numpy
scikit-learn
xgboost
sentence-transformers
faiss-cpu
pydantic>=2
openai>=1.35
joblib
scipy
pyarrow
typing_extensions